requirements.txt
ninja
numpy==1.23.0
tqdm==4.49.0
torchtext==0.13.0
GitPython==3.1.30
requests==2.23.0
fasttext==0.9.1
nltk==3.6.6
editdistance==0.5.3
transformers==4.10.1
sklearn==0.0
omegaconf==2.1.0
lmdb==0.98
termcolor==1.1.0
iopath==0.1.8
datasets==1.2.1
matplotlib
pycocotools==2.0.2
ftfy==5.8
pytorch-lightning==1.6.0
psutil
pillow==9.3.0
sentencepiece
timm
h5py
fvcore
fairscale
opencv-python
git+https://github.com/cocodataset/panopticapi.git
